Pre-commit hook lints all JS files.
The auto-lint.py file should be installed as a pre-commit hook
with the name "pre-commit" (no quotes) in the .git/hooks directory
of every repository that wishes to use a linter. It assumes the
existence of a lint.sh file in the /usr/bin directory. If this
script does not exist, it will output a warning but will not
prevent commits.

Currently, only JavaScript files are linted. Other file's extensions
are printed to the console as a proof of concept that this hook
can distinguish which linters to use on which files.

In truth, the linter in the current state shouldn't be used.
Further issues need to be completed before it should be used in
production.