Lint only files modified in a commit.
Only files that have been added (A), modified (M) copied (C),
or broken* (B), will be passed to the linter. Currently, it
only handles JavaScript but the pattern will still apply when
we support more languages. The context of these
modifications/additions/copies are within the current commit
being proposed. Thus, only work you are making changes to
(which are tracked by git) will be linted.

*The Broken git file type indicates that the majority of a
file has changed and it is more effective to consider it an
entire file remove + add with the same name than it is to
do a line-by-line set of changes. Fun fact for the day!
Aren't you glad you read long commit messages?