Linting diffs used in pre-commit hook.
When a file is modified relative to a previous git revision,
the pre-commit hook (currently titled auto-lint.py) will compare
the difference between the linting outputs of the file in the
previous commit versus the file you are currently attempting to
commit. This difference will be noted and logged to a file.

If the linting output is different it probably either
indicates more errors or fewer. If more errors are produced,
this is a sign that the commit should be amended to address them.
If fewer are produced, that is a good thing. A followup issue
will address the correction of previous errors/warnings.

This issue requires that the lint.sh used does not exit with
status code 2 (which another issue addresses) as that would
prevent the commit from succeeding. This currently happens when
a JSHint check fails to pass.