Newly added JS files are linted.
When a file is added, it must have no linting errors or a log
and console message warning will be produced.

Additionally, the return value of lint.sh will no longer always
be 0. Previously we had to do this because we weren't catching
the subprocess.CalledProcessError that resulted from a non-0
return value. However, it turns out we can both catch this error
as well as get the output from the script we execute. The
behavior now will be to output an exit code of 0 if and only if
neither of our linters reported any warnings/errors.