Smaller lintdiffs are celebrated.
Previously the linting tools could not distinguish between removing
linting errors and introducing new errors. Both were considered
worthy of creating a log file and printing out a notice to the
console. Thanks to the magic of regular expressions, we have the
capability to distinguish between newly introduced errors and
everything else, so the tool no longer whines when better lintdiffs
are produced by developer changes.