Fixed scoping issue for comparison func
Previously a refactor had moved compareFullMessages
out of another function to an outer layer of scope.
However, this change made the variable filename become
undefined when that function was evaluated within the
sort call.

The Javascript pattern used to resolve this
is to rebind the "this" reference and insert a default
argument to cause function currying such that filename
becomes visible and defined while maintaining the
function signature sort() requires.