Introduce setup.py build system
(This commit was added later to simplify the history before open-sourcing
this repository.)

Introduces a Python build system using setup.py and the setuptools and
pkg_resources packages. We move the auto-lint.py script to its own
"difflint" package, and make it into a python module instead of a script.
We autogenerate a script using setuptools that calls the main() entry
point in difflint.

All the data files, including lint.sh and the jshintrc files, go into a
data/ subdirectory of the difflint package, from where they can be
accessed with pkg_resources, no matter where they are installed.

We leave the example .jscsrc and .jshintrc files in the repository root
because they are meant to be copied to your home directory.