Move "check" functionality to main script
This moves the functionality of checking for the external linter
executables from the lint.sh script into Python code. It simplifies the
code quite a bit and is part of the effort not to have code in two
languages (Python and shell.)

We don't need to check for the existence of the lint.sh script or the
auxiliary files, because due to pkg_resources and setup.py they are now
guaranteed to be included in the Python package.

The use of shutil.which() requires Python 3.3.