Remove lint.sh script altogether
This ports the lint.sh script to Python code, which is more portable in
the context of our Python package since it doesn't rely on sh as an
external dependency.

Moves the LintOutput object to lint.py and adds a run_command() method
that simplifies checking the output and exit code of the external linting
tool.