Support linting Python files
This adds linting for Python files with the PEP8 and PyFlakes packages.
Instead of launching subprocesses for them, it's convenient enough to
run them through their Python API, and this makes it extra easy to
generate output in the required format. (PyFlakes doesn't support custom
reporters from the command line, as far as I can tell.)

This requires splitting LintOutput out into its own file, because
otherwise it would be a circular dependency.