Consistently use an exit code
Make sure that an integer exit code is always returned. For the error
exit code for missing files, we would like to use os.EX_OSFILE, but that
is not available on all platforms, so we just use its numerical value.
(On other platforms that value should be meaningless, so that's OK.)