Linters are optional now
Previously we required the Javascript linters jscs and
jshint and would abort the program if they were not found.
We also tightly coupled our hardcoded file extensions to
the linters we supported, limiting situations where someone
wanted to use a linter for a file extension we did not
anticipate.

The biggest victory, however, is our newfound ability to
enable or disable supported linters via the new
enabled_linters.json file. As of this commit, the file will
be installed in the data folder of difflint but in the future
we will want to support user-customizable overrides in their
repository root, following the same JSON format.

One casualty of decoupling the linters from file extensions
is that we are not able to easily support various LintOutput
classes. This forced me to move the logic for the python
linters to the lint.py file and rename the lint_python.py file
to reflect the fact that it only contains reporters now.

Another modification was to enhance the functionality of the
--check mode and the automatic checking for linters at runtime
in the normal execution mode. They now will both use the
configuration file to determine which linters are required.