Cleaned up imports.
I like the pattern of using the full package/function
name for functions outside of the difflint files. I will
try to stick with this pattern unless the fully qualified
name is too friggin' long, in the which case I'll make
exceptions.

(See lint.py which has a from x import y that I'm keeping.)