Fixed renaming bug.
Moving files or renaming them caused the linter to
erroneously believe they had different linting output
because the filename was included in the linting output.

This was removed as it provided no benefit. The files are
compared to older versions of themselves before they are
added to the difflint output, so we already avoided the
concern of two files introducing and removing the same
error not being detected.  We gained nothing by putting
the filename in the autogenerated output.

This required removal from the reporters and a small
regex change.  I'm keeping the leading "|" character
as I find it makes the output slightly more readable.